requirements.txt
boto3
pandas
openpyxl
delta-spark
